With this commit, we add lazy-loading.
We do it using the magic prototypal method "Object.defineProperty".
"defineProperty" adds a field to the argument, kind of like the assignment
operator. However, unlike the assignment operator, this method gives
fine-grained, low-level control over how the runtime accesses the field.
For example, you can determine if the field is mutable or immutable, whether
or not its visible via certain APIs. The key feature that allows us to lazy-
load is the `get` property. When defined, you can specify a callback function
which triggers on property access.

As an example, support we want to import a module named "bar".
Using defineProperty, we can make it such that a client writing the expression
"exports.bar" will call a function we provide to return the value of the
expression "exports.bar".

Now, we specify that our callback function will load the module, knowing
that NodeJS will cache the module so multiple field accesses don't
repeat work.